requirements.txt
   fastapi
   uvicorn
   requests
   python-dotenv
   langchain
   langchain-openai
   langchain-community
